refactor: Ladder 도메인 수정 및 테스트 추가
